+++ lib/curl_fopen.h
@@ -24,7 +24,7 @@
  *
  ***************************************************************************/
 
-#include "curlx/fopenx.h"
+#include "curlx/fopen.h"
 
 CURLcode Curl_fopen(struct Curl_easy *data, const char *filename,
                     FILE **fh, char **tempname);
