+++ lib/vtls/keylog.c
@@ -33,7 +33,7 @@
 #include "keylog.h"
 #include <curl/curl.h>
 #include "../escape.h"
-#include "../curlx/fopenx.h"
+#include "../curlx/fopen.h"
 
 /* The last #include files should be: */
 #include "../curl_memory.h"
